Create Mapper

+++ backend/src/main/java/com/net1707/backend/model/QuizResult.java
@@ -14,7 +14,7 @@ public class QuizResult {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer quizResultId;
+    private Long quizResultId;
 
     @ManyToOne
     @JoinColumn(name = "customerId", nullable = false)
