+++ backend/src/main/java/com/net1707/backend/repository/StaffRepository.java
@@ -5,8 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-
 @Repository
-public interface StaffRepository extends JpaRepository<Staff,Integer>{
+public interface StaffRepository extends JpaRepository<Staff,Long>{
     Optional<Staff> findByEmail(String email);
 }
